Flatcar: enable updates from the AKS group
Automatic atomic updates are a key element of Flatcar and we should try
to have them on by default. This requires FLUO or kured as reboot
coordinators on Kubernetes, and a way to control which version to
follow for the worker nodes.
Use a custom AKS group in the update service to specify the Flatcar
version, and enable update-engine. FLUO or kured should be deployed by
the RP.